requirements.txt
fastf1==3.4.0
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
groq==0.11.0
python-dotenv==1.0.1
scipy==1.14.1
requests==2.32.3
urllib3==2.2.3
